cloud_run/requirements.txt
apache-beam[gcp]
oauthlib
requests
requests-oauthlib
google-cloud-storage
google-auth
google-auth-oauthlib
Flask